jf: kernel: merge I9505GUEUDNL3
Conflicts:
	Makefile
	arch/arm/kernel/fiq.c
	arch/arm/mach-msm/Kconfig
	arch/arm/mach-msm/acpuclock-krait.c
	arch/arm/mach-msm/board-8064.h
	arch/arm/mach-msm/hotplug.c
	arch/arm/mach-msm/mdm_common.c
	arch/arm/mach-msm/timer.c
	arch/arm/mm/dma-mapping.c
	arch/arm/mm/mmu.c
	drivers/hid/hid-logitech-dj.c
	drivers/hid/hid-picolcd.c
	drivers/media/radio/radio-iris.c
	drivers/media/video/msm/vfe/msm_vfe32.c
	drivers/misc/qseecom.c
	drivers/sensorhub/ssp_firmware.c
	drivers/video/au1100fb.c
	drivers/video/au1200fb.c
	drivers/video/msm/mdp.c
	drivers/video/msm/mdp4.h
	drivers/video/msm/mdp4_overlay.c
	drivers/video/msm/mdp4_overlay_dsi_video.c
	drivers/video/msm/mdp4_overlay_writeback.c
	drivers/video/msm/mhl_v2/sii8240.c
	drivers/video/msm/mipi_dsi.c
	drivers/video/msm/msm_fb.c
	drivers/video/msm/vidc/1080p/ddl/vcd_ddl_shared_mem.c
	drivers/video/msm/vidc/1080p/ddl/vcd_ddl_shared_mem.h
	drivers/video/msm/vidc/1080p/ddl/vcd_ddl_vidc.c
	firmware/ssp_rev03.fw.ihex
	fs/proc/proc_avc.c
	include/net/tcp.h
	kernel/events/core.c
	kernel/sched/core.c
	kernel/sched/sched.h
	kernel/time/timekeeping.c
	mm/vmscan.c
	net/ipv4/ping.c
	net/ipv4/syncookies.c
	net/ipv6/route.c
	net/ipv6/tcp_ipv6.c
	sound/usb/card.c

+++ kernel/power/wakelock.c
@@ -82,7 +82,7 @@ int get_expired_time(struct wake_lock *lock, ktime_t *expire_time)
 		return 0;
 	jiffies_to_timespec(-timeout, &delta);
 	set_normalized_timespec(&ts, kt.tv_sec + tomono.tv_sec - delta.tv_sec,
-				kt.tv_nsec + tomono.tv_nsec - delta.tv_nsec);
+				(s64)kt.tv_nsec + tomono.tv_nsec - delta.tv_nsec);
 	*expire_time = timespec_to_ktime(ts);
 	return 1;
 }
